update API in preparation for format changes
This changes the internal representation so that
we don't use the binary encoding directly.

We also change the API so that third party caveat
ids and the macaroon id itself are []byte not string,
and relax the length constraints which will only
apply when marshaling to the old binary format.