fix unmarshaling for Go tip
The Go compiler has been fixed so that unexported
embedded struct fields cannot be assigned to.
This fix makes macaroon JSON unmarshaling work
with that.

This commit is made to an early version of v2-unstable
that some repos depend on that can't yet move to using the
latest v2-unstable or v2.